backend/pyproject.toml
[project]
name = "tutorial"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.123.7",
    "langchain[openai]>=1.1.0",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.1",
    "uvicorn>=0.38.0",
    "yfinance>=0.2.66",
    "scikit-learn>=1.6.0",
    "numpy>=1.26.0",
]
